+++ src/shared/types/general.ts
@@ -5,7 +5,6 @@ type TopFilters = {
   tags: string[];
   objectTypeRegistry: string[];
   dates: string[];
-  exploitation: string[];
 };
 
 enum COLUMN_TYPE {
@@ -180,6 +179,9 @@ type Row = ArtificialRow & {
   operational_monitoring: string | null;
   analytical_monitoring: string | null;
   bpmn_key: string | null;
+  output_table: string | null;
+  allocation_assessment_class: string | null;
+  allocation_assessment_parameters: string | null;
 };
 
 type ColumnsFilter = Record<keyof Row, Array<string>>;